+++ infrastructures/doctrine/odm/AbstractStandardDocument.php
@@ -30,7 +30,6 @@ use Teknoo\States\Proxy\ProxyInterface;
 /**
  * Default Stated class implementation with a doctrine document.
  *
- *
  * @copyright   Copyright (c) 2009-2021 EIRL Richard Déloge (richarddeloge@gmail.com)
  * @copyright   Copyright (c) 2020-2021 SASU Teknoo Software (https://teknoo.software)
  *
@@ -38,6 +37,8 @@ use Teknoo\States\Proxy\ProxyInterface;
  *
  * @license     http://teknoo.software/license/mit         MIT License
  * @author      Richard Déloge <richarddeloge@gmail.com>
+ *
+ * @deprecated Use Teknoo\States\Doctrine\AbstractStandardObject instead
  */
 abstract class AbstractStandardDocument implements ProxyInterface
 {